requirements.txt
pyserial~=3.5
crccheck~=1.2.0
cryptography~=38.0.1
